Revert type change

+++ module/strategy/StrategiseLook/src/StrategiseLook.cpp
@@ -29,9 +29,9 @@ namespace module::strategy {
             // Use configuration here from file StrategiseLook.yaml
             this->log_level         = config["log_level"].as<NUClear::LogLevel>();
             cfg.ball_search_timeout = duration_cast<NUClear::clock::duration>(
-                std::chrono::duration<float>(config["ball_search_timeout"].as<float>()));
+                std::chrono::duration<double>(config["ball_search_timeout"].as<double>()));
             cfg.goal_search_timeout = duration_cast<NUClear::clock::duration>(
-                std::chrono::duration<float>(config["goal_search_timeout"].as<float>()));
+                std::chrono::duration<double>(config["goal_search_timeout"].as<double>()));
         });
 
         // Trigger on FilteredBall to update readings
@@ -40,7 +40,7 @@ namespace module::strategy {
             [this](const FilteredBall& ball) {
                 // If we have a ball and it is recent, look at it
                 if (NUClear::clock::now() - ball.time_of_measurement < cfg.ball_search_timeout) {
-                    emit<Task>(std::make_unique<Look>(ball.rBCt.cast<float>(), true));
+                    emit<Task>(std::make_unique<Look>(ball.rBCt.cast<double>(), true));
                 }
             });
 
@@ -51,7 +51,7 @@ namespace module::strategy {
                 // If we have goals, with at least one measurement and the goals are recent, look at the goals
                 if (!goals.goals.empty() && (NUClear::clock::now() - goals.timestamp < cfg.goal_search_timeout)) {
                     // Convert goal measurement to cartesian coordinates
-                    Eigen::Vector3d rGCc = sphericalToCartesian(goals.goals[0].measurements[0].srGCc.cast<float>());
+                    Eigen::Vector3d rGCc = sphericalToCartesian(goals.goals[0].measurements[0].srGCc.cast<double>());
                     // Convert to torso space
                     Eigen::Vector3d rGCt = Eigen::Isometry3d(sensors.Htw * goals.Hcw.inverse()).rotation() * rGCc;
                     // Look at the goal
